+++ R/praatScriptPitch.R
@@ -25,6 +25,21 @@
 #' @param gender.attribute Name of the LaBB-CAT participant attribute that contains the
 #'     participant's gender - normally this is "participant_gender".
 #' @param value.for.male The value that the gender.attribute has when the participant is male.
+#' @param sample.points A vector of numbers (0 <= sample.points <= 1) specifying multiple
+#'     points at which to take the measurement.  The default is NULL, meaning no
+#'     individual measurements will be taken (only the aggregate values identified by
+#'     get.mean, get.minimum, and get.maximum).  A single point at 0.5 means one
+#'     measurement will be taken halfway through the target interval.  If, for example, 
+#'     you wanted eleven measurements evenly spaced throughout the interval, you would
+#'     specify sample.points as being 
+#'     c(0.0, 0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 1.0).  
+#' @param interpolation If sample.points are specified, this is the interpolation to use
+#'     when getting individual values. Possible values are 'nearest' or 'linear'.
+#' @param skip.errors Sometimes, for some segments, Praat fails to create a Pitch
+#'     object. If skip.errors = TRUE, analysis those segments will be skipped, and corresponding
+#'     pitch values will be returned as "--undefined--". If skip.errors = FALSE, the error
+#'     message from Praat will be returned in the Error field, but no pitch measures will
+#'     be returned for any segments in the same recording.
 #' @return A script fragment which can be passed as the praat.script parameter of
 #'     \link{processWithPraat} 
 #' 
@@ -49,7 +64,7 @@
 #' }
 #' @keywords praat
 #' 
-praatScriptPitch <- function(get.mean = TRUE, get.minimum = FALSE, get.maximum = FALSE, time.step = 0.0, pitch.floor = 60, max.number.of.candidates = 15, very.accurate = FALSE, silence.threshold = 0.03, voicing.threshold = 0.5, octave.cost = 0.01, octave.jump.cost = 0.35, voiced.unvoiced.cost = 0.35, pitch.ceiling = 500, pitch.floor.male = 30, voicing.threshold.male = 0.4, pitch.ceiling.male = 250, gender.attribute = 'participant_gender', value.for.male = "M") {
+praatScriptPitch <- function(get.mean = TRUE, get.minimum = FALSE, get.maximum = FALSE, time.step = 0.0, pitch.floor = 60, max.number.of.candidates = 15, very.accurate = FALSE, silence.threshold = 0.03, voicing.threshold = 0.5, octave.cost = 0.01, octave.jump.cost = 0.35, voiced.unvoiced.cost = 0.35, pitch.ceiling = 500, pitch.floor.male = 30, voicing.threshold.male = 0.4, pitch.ceiling.male = 250, gender.attribute = 'participant_gender', value.for.male = "M", sample.points = NULL, interpolation = 'linear', skip.errors = TRUE) {
     script <- paste(
         "\npitchfloor = ", pitch.floor,
         "\nvoicingthreshold = ", voicing.threshold,
@@ -88,26 +103,80 @@ praatScriptPitch <- function(get.mean = TRUE, get.minimum = FALSE, get.maximum =
     }
     script <- paste( # ensure the sound sample is selected
         script, "\nselect Sound 'sampleName$'", sep="")
+    if (skip.errors) { ## use nocheck
+        script <- paste(
+            script,
+            "\n# nocheck to prevent the whole script from failing, and then check for object after",
+            "\nnocheck ", sep="")
+    } else {
+        script <- paste(script, "\n", sep="")
+    }
+    
     script <- paste(
-        script, "\nTo Pitch (ac): ", time.step, ", pitchfloor, ", max.number.of.candidates, ", ",
+        script,
+        "To Pitch (ac): ", time.step, ", pitchfloor, ", max.number.of.candidates, ", ",
         "veryaccurate$, ", silence.threshold, ", voicingthreshold, ", octave.cost, ", ",
         octave.jump.cost, ", ", voiced.unvoiced.cost, ", pitchceiling", sep="")
+    
+    script <- paste(
+        script,
+        "\n# check that a Pitch object was created",
+        "\nobjectCreated = extractWord$(selected$(), \"\") = \"Pitch\"", sep="")
     if (get.mean) {
         script <- paste(script,
-                       "\nmeanPitch = Get mean: targetStart, targetEnd, \"Hertz\"",
+                       "\nif objectCreated",
+                       "\n  meanPitch = Get mean: targetStart, targetEnd, \"Hertz\"",
+                       "\nelse",
+                       "\n  meanPitch = 1/0", # --undefined--
+                       "\nendif",
                        "\nprint 'meanPitch' 'newline$'", sep="")
     }
     if (get.minimum) {
         script <- paste(script,
-                       "\nminPitch = Get minimum: targetStart, targetEnd, \"Hertz\", \"Parabolic\"",
+                       "\nif objectCreated",
+                       "\n  minPitch = Get minimum: targetStart, targetEnd, \"Hertz\", \"Parabolic\"",
+                       "\nelse",
+                       "\n  minPitch = 1/0", # --undefined--
+                       "\nendif",
                        "\nprint 'minPitch' 'newline$'", sep="")
     }
     if (get.maximum) {
         script <- paste(script,
-                       "\nmaxPitch = Get maximum: targetStart, targetEnd, \"Hertz\", \"Parabolic\"",
+                       "\nif objectCreated",
+                       "\n  maxPitch = Get maximum: targetStart, targetEnd, \"Hertz\", \"Parabolic\"",
+                       "\nelse",
+                       "\n  maxPitch = 1/0", # --undefined--
+                       "\nendif",
                        "\nprint 'maxPitch' 'newline$'", sep="")
     }
+    if (!is.null(sample.points)) {
+        for (point in sample.points) {
+            varname = paste("time_", stringr::str_replace(point, "\\.","_"), "_for_pitch", sep="")
+            ## first output absolute point offset
+            script <- paste(script, "\npointoffset =",
+                            " targetAbsoluteStart + ", point, " * targetDuration", sep="")
+            script <- paste(script, "\n", varname, " = pointoffset", sep="")
+            script <- paste(script, "\nprint '", varname, "' 'newline$'", sep="")
+            ## now use the relative point offset
+            script <- paste(script, "\npointoffset =",
+                            " targetStart + ", point, " * targetDuration", sep="")
+            varname = paste("pitch_time_", stringr::str_replace(point, "\\.","_"), sep="")
+            script <- paste(script,
+                            "\nif objectCreated",
+                            "\n  ", varname,
+                            " = Get value at time: pointoffset, \"Hertz\", \"",interpolation,"\"",
+                            "\nelse",
+                            "\n  ", varname, " = 1/0", # --undefined--
+                            "\nendif",
+                           sep="")
+            script <- paste(script, "\nprint '", varname, ":0' 'newline$'", sep="")
+        } ## next sample point
+    }
     script <- paste( # remove pitch object
-        script, "\nRemove\n", sep="")
+        script,
+        "\nif objectCreated",
+        "\n  Remove",
+        "\nendif\n", sep="")
+        
     return(script)
 }